backend/requirements.txt
asgiref==3.8.1
boto3==1.40.6
botocore==1.40.6
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==4.1
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
runscript==0.3.5
s3transfer==0.13.1
setproctitle==1.3.6
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
